0901 알고리즘

+++ src/baekjoon/b9461/Main.java
@@ -6,14 +6,14 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		int N = sc.nextInt();
-		int[] arr = new int[101];
+		long[] arr = new long[101];
 		arr[1] = 1;
 		arr[2] = 1;
 		arr[3] = 1;
 		for (int i = 4; i <= 100; i++) {
 			arr[i] = arr[i-2] + arr[i-3];
 		}
-		int[] answer = new int[N];
+		long[] answer = new long[N];
 		for (int i = 0; i < N; i++) {
 			int a = sc.nextInt();
 			answer[i] = arr[a];
